(include file fix) for solving compile .c==>.o bug via arm-none-eabi-gcc during c-hacks construction
